feat: Update npm dependencies and package versions
This commit updates the npm dependencies and package versions in the project. The "@expo/metro-runtime" dependency is updated to version "~3.2.1" and the "expo-status-bar" dependency is updated to version "~1.12.1". Additionally, the "react-native" dependency is updated to version "^0.74.1" and the "react-native-gesture-handler" dependency is updated to version "~2.16.1". The "react-native-reanimated" dependency is updated to version "~3.10.1" and the "react-native-safe-area-context" dependency is updated to version "^4.10.1". Finally, the "react-native-screens" dependency is updated to version "^3.31.1". These updates ensure that the project is using the latest stable versions of the packages, improving performance and compatibility.