feat: Enable authentication check in Home component
This commit enables the authentication check in the Home component by implementing token verification using the "jsonwebtoken" library. If the user is not authenticated, a message indicating that the user is not authenticated will be displayed. This improves the security of the app by ensuring that only authenticated users can access the Home screen.