feat: Fix issue with playing next song in PlayerPlaylist screen
This commit fixes a bug in the PlayerPlaylist screen where the next song was not playing correctly. The code changes in the index.jsx file update the logic for playing the next song when reaching the end of the playlist. Now, when the current song is the last one, it will go back to the first song in the playlist.

Closes #123