style: Update login button color and style
This commit updates the color and style of the login button in the SignIn screen. The code changes in the styles.js file modify the loginButton style to have an underline and a color of "#555555". This enhances the visual appearance of the button and improves the overall design of the SignIn screen.