feat: Update Library screen UI
This commit updates the UI of the Library screen. The code changes in the styles.js file add new styles for the cardContainer, cardImage, cardTextContainer, cardText, artistText, addToQueueButton, addToQueueButtonText, and musicPlayButton. These changes enhance the visual appearance of the music cards and improve the overall design of the Library screen.

Closes #123