feat: Update SignUp screen layout and functionality
The code changes in the SignUp screen improve the layout and functionality of the form. The exit button is moved inside the containerUser view, ensuring it is displayed correctly. The form inputs are rearranged and the password input now has a toggle button to show or hide the password. Additionally, the confirm password input is added for better user experience. These changes enhance the usability and visual appearance of the SignUp screen.